fix: Adjust duration display and timeout settings across components
This commit updates the duration display in the RunDetailsModal and RunTable components to round the duration to one decimal place. Additionally, the timeout for the ScenariosPage has been increased from 2 seconds to 20 seconds to improve loading performance. These changes enhance the user experience by providing clearer information and more reliable loading times.

+++ usefly/models/agent_run.py
@@ -44,7 +44,7 @@ class PersonaRunCreate(BaseModel):
     persona_type: str
     is_done: bool = False
     timestamp: datetime
-    duration_seconds: Optional[int] = None
+    duration_seconds: Optional[float] = None
     platform: str = "web"
     location: Optional[str] = None
     error_type: Optional[str] = None
@@ -65,7 +65,7 @@ class PersonaRunResponse(BaseModel):
     persona_type: str
     is_done: bool
     timestamp: datetime
-    duration_seconds: Optional[int]
+    duration_seconds: Optional[float]
     platform: str
     location: Optional[str]
     error_type: Optional[str]
